refactor : refactor the useScanner hook
