+++ frontend/src/routes/PrivateRoute.tsx
@@ -1,13 +1,13 @@
 import { Routes, Route, Outlet } from 'react-router-dom'
 import Profile from '../pages/Profile'
-import ScanQRCode from '../pages/TransportWorker/ScanQRCode'
+import ScanQRCode from '../pages/TransportWorker/Scan'
 import Cart from '../pages/TransportWorker/Cart'
 import UnloadSuccess from '../pages/Success'
 import { useContext, useEffect } from 'react'
 import { AuthContext } from '../contexts/auth'
 import Dashboard from '../pages/Dashboard'
 import TaskDetailPage from '../pages/TransportWorker/TaskDetailPage'
-import PickerScanPage from '../pages/Picker/ScanBin'
+import PickerScanPage from '../pages/Picker/Scan'
 import CreateTaskPage from '../pages/Picker/CreateTask'
 
 import { TransportWorkCartProvider } from '../contexts/cart'
